move RefcountDB into db
