+++ ethereum/db/__init__.py
@@ -1,3 +1,4 @@
 from .base_db import BaseDB
 from .ephem_db import EphemDB
 from .overlay_db import OverlayDB
+from .refcount_db import RefcountDB
